Minor refactoring:
* Renames "database-operations" -> "db"
* Cleans up deleteItem
* Renames "deserializeFromRequest" -> "deserializeRequest"
* Changes wording of error messages
* Adds additional type checks for RESTPathGenerator
* Adds Total Count header to getCount operation
* Freezes constant for your protection ;)
