+++ src/api/db/deleteItems.js
