+++ src/api/db/getShallow.js
